+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 * Add masquerading to allow admins to impersonate other users ([#147](https://gitlab.com/yale-sdmp/vesta/issues/147)).
 * Add information about pending group requests to user dashboard ([#696](https://gitlab.com/yale-sdmp/vesta/issues/696)).
 * Add building abbreviations ([#570](https://gitlab.com/yale-sdmp/vesta/issues/570)).
+* Assign Suites button still visible when no possible suites to assign ([#976](https://gitlab.com/yale-sdmp/vesta/issues/976)).
 
 ### Changed
 * Change form cancel operation to standardize redirects to page referrer ([#979](https://gitlab.com/yale-sdmp/vesta/issues/979)).